+++ ipfs/Cargo.toml
@@ -1,0 +1,16 @@
+[workspace]
+resolver = "2"
+members = [
+	"effector",
+	"types",
+	"cid"
+]
+[workspace.dependencies]
+ipfs-effector-types = { path = "types" }
+ipfs-effector = { path = "effector" }
+ipfs-cid = { path = "cid" }
+
+marine-rs-sdk = { version = "0.14.0", features = ["logger"] }
+marine-rs-sdk-test = "0.16.0"
+
+eyre = "0.6.12"
