+++ ipfs/cid/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ipfs-effector-cid"
+version = "0.1.0"
+edition = "2021"
+build = "built.rs"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[dependencies]
+
+[build-dependencies]
+built = "0.7.0"
