+++ ipfs/effector/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ipfs-effector"
+version = "0.1.0"
+edition = "2018"
+
+[dependencies]
+eyre = "0.6.12"
+itertools = "0.12.1"
+log = "0.4.20"
+marine-rs-sdk = { workspace = true }
+ipfs-effector-types = { workspace = true}
+
+[dev-dependencies]
+marine-rs-sdk-test = { workspace = true }
