+++ ipfs/effector/module.yaml
